Rename cache to symbol scope
