+++ global.json
@@ -1,0 +1,7 @@
+{
+  "sdk": {
+    "version": "7.0.202",
+    "rollForward": "latestMajor",
+    "allowPrerelease": true
+  }
+}
